Adjust install/varnish.params to use the default thread pool settings.
By mistake, a version of this file was committed that limited the number of threads in each thread to less than what the default setting would be. The thread pool settings in varnish.params now only speciies the number of threadpools to 4 rather than the default 2 to decrease lock contention and provide more threads.